Added rebalancing of infantry on addition and removal.
When units are added, infantry are checked to see if there are the right number of supported units. Same on removal. Added unit tests with a bit of a terrible hack for now to see the internal state of the army. Can be removed later, but it works well enough for now